#138 bug fix: WinRate is no longer an int

+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/UserGetDTO.java
@@ -20,7 +20,7 @@ public class UserGetDTO {
   private double winRate;
 
   public double getWinRate() {return winRate;}
-  public void setWinRate(int winRate) {this.winRate = winRate;}
+  public void setWinRate(double winRate) {this.winRate = winRate;}
 
   public int getLengthPR() {return lengthPR;}
   public void setLengthPR(int lengthPR) {this.lengthPR = lengthPR;}
